server: config-management API + settings UI + transcode trigger
Re-applies the config/settings slice from the parked `devel` WIP onto current
main, deliberately omitting devel's catalog.rs rewrite (which regressed the
BackgroundRuntime SIGTERM fix + ANN index + refine_factor on main).

- GET/POST /api/config: read and update config in the browser, persisted to
  config.toml; POST validates that library paths exist and are directories.
- POST /api/transcode: dry-run (eligibility count) and live transcode from the
  web, reporting progress.
- Config-driven scan/ingest: /api/scan and /api/ingest fall back to library
  paths from config ("Add paths in Settings") instead of requiring a path arg.
- Settings web-UI page + styles; AppState carries Arc<RwLock<Config>> +
  config_path; serve loads config.toml at startup.
- core::Config gains merge()/to_file() (with tests); SearchEngine derives Clone.
- New integration tests for the config API handlers (catalogy-server had zero
  handler-level coverage before).

+++ crates/catalogy-extract/src/worker.rs
@@ -34,7 +34,7 @@ fn strategy_from_config(config: &ExtractionConfig) -> ExtractionStrategy {
             seconds: config.frame_interval_seconds,
         },
         _ => ExtractionStrategy::Adaptive {
-            scene_threshold: config.scene_threshold,
+            scene_threshold: config.scene_threshold as f32,
             max_interval_seconds: config.max_interval_seconds,
         },
     }
@@ -302,7 +302,7 @@ mod tests {
                 scene_threshold,
                 max_interval_seconds,
             } => {
-                assert!((scene_threshold - 0.4).abs() < f32::EPSILON);
+                assert!((scene_threshold - 0.4_f32).abs() < f32::EPSILON);
                 assert_eq!(max_interval_seconds, 30);
             }
             _ => panic!("Expected Adaptive strategy"),
